fix: fixed small naming issues
