+++ app/entities/SudokuRegistry.py
@@ -6,7 +6,7 @@ class SudokuRegistry(Base):
   __tablename__ = "sudoku_registry"
 
   id = Column(UUID, primary_key=True, index=True)
-  puzzle_id = Column(UUID, nullable=False)
+  sudoku_id = Column(UUID, nullable=False)
   user_id = Column(UUID, nullable=False)
   solving_time = Column(Float, nullable=False)
   is_applicable = Column(Boolean, nullable=False)
