+++ app/repositories/__init__.py
@@ -1,3 +1,3 @@
 from .UserRepository import UserRepository
 from .SudokuRepository import SudokuRepository
-from .SudokuEntryRepository import SudokuEntryRepository
+from .SudokuRegistryRepository import SudokuRegistryRepository
